Restored most of Calendar's functionality.
I kept some of the PHP code for reference, which was used with my own
parallel programming framework. I'm not quite sure if and how I will use
Threads in Java; in the end, there were only two threads in PHP, Input and
Calendar. In the end it is questionable if threads are actually needed.