Refactored TerminalWidget and TerminalWidgetAbstract
There are Widgets which are supposed to have generic InputObservers, while
others are more specialized like WidgetYesNo or Calendar and will react
more directly to input, either forwarding it to their own observers like
WidgetYesNo or handle it directly like the current (debugging) speed
increase for Calendar.

Refactoring only went so far to restore everything to working order as it
was before, next step will be to refactor the implementations/extended
classes properly.