Mock tilling of tiles using no tool.
Had to do some changes. Room gets a refresh method that will iterate over
tiles and recreate the groups for WASDSelect. This is necessary so that
WASDSelect is up to date.
WASDSelectObserver management is passed through WASDSelect by room.
Room's WASDSelect works with WASD again the inventory works with cursor keys.
Game has the WASDSelect observer methods and do the tilling for now without
checking for the proper tool.