Changed approach to handling crops.
Using a generic approach - ie have 'Crop' instantiate 'Wheat' - as it is
possible in PHP would be a major pain in the ...neck... for Java. But in the
end it is not the best of all designs anyway, as code changes are necessary
for each new/modified crop.
So there is a class which essentially contains a list of all crops with their
properties and a unique ID.
Having some sort of unique ID is necessary anyway in the long run if there is to
be some sort of persistence.