Added ImmutableTile, Interactive items and Bed.
ImmutableTiles cannot be changed by tools and are supposed to hold items which
the player should not change and which are to be used for basic game
functionality, such as a bed, a stove, exits and so on.

Interactive items are items which may change many aspects of a game. A bed will
forward time, a shop will manipulate player inventory/money, an exit will move
the player to another room and so on.