Added class 'rooms' which will keep a collection of all available rooms
Also added Exit, which allows to go from one room to the other, and added
it to Farm and the new room YourHouse.
I would prefer to use an instance of Room for Exit as target, in order to use
its getName for the exit, but this would need dependency injection of Rooms into
a single room or having Rooms as a singleton or whatever. Having a name which
may be distinct to the room name also allows for more flexibility in naming
exits.