Make time pass on tool use.
With no animations (and no Cooldown), it is basically possible to run through
a day's energy within a few seconds. Adding a cooldown limits the amount of
work that can be done within a real life time frame, but is artificial in
the end and should not be too long. Fast forwarding time seems to be a good
compromise between allowing players to get things done while not allowing them
to be done at 06:20 with everything.