Incomplete ItemFactory, which is supposed to be able to create all items.
This is for persistence: every item is supposed to have a unique ID, which will
be stored when saving the game. If the game is loaded again, the Item can be
reinstantiated again using the ID and the ItemFactory as well as any additonal
data.

I'm uncertain on how to handle different variations of the same object, such
as Hoe and Hoe+1 (higher quality hoe). One approach is to create a prototype
for every variety (HOE, HOE_COPPER) or to store the HOE's variety in the
savegame, which would result in larger save files as well as a more complicated
format with variable length items - wood would only have one additional short
for amount, while crops would have additional data for age and so on.